Use native Net::FTP rather than double-bags for FTPS.
It seems as if Net::FTP added TLS support as of ruby 2.4,
which conflicts with double-bags. This switches ftps from
using double-bags to using the native ruby Net::FTP.

I needed support for ascii file transfers, so this also
adds support for that. Luckily this was pretty easy to do.

Warning. This might break implicit ftps. I don't have any
way to test that functionality.

Also, I'm not a ruby programmer. I simply needed this
working in my environment. This could likely be cleaned
up quite a bit.